+++ neuromapp/keyvalue/CMakeLists.txt
@@ -29,8 +29,7 @@ add_executable(MPI_Exec_kv utils/statistic.cpp mpiexec.cpp )
 
 
 # Adding MPI_LIBRARIES adds also the -Bdynamic flag, which makes execution crash on BG/Q
-target_link_libraries (MPI_Exec_kv map_store)
-#target_link_libraries (MPI_Exec_kv map_store ${MPI_LIBRARIES})
+target_link_libraries (MPI_Exec_kv map_store ${MPI_CXX_LIBRARIES})
 
 
 if(IS_DIRECTORY ${SKV_PATH})
