+++ neuromapp/keyvalue/main.cpp
@@ -139,7 +139,7 @@ void keyvalue_content(po::variables_map const& vm){
     std::stringstream command;
 
     command << "OMP_NUM_THREADS=" << nt << " mpirun -n " << np
-            << " /Users/ewart/Documents/miniapps/b/neuromapp/keyvalue/MPI_Exec_kv  -b " << backend
+            << " /Users/ewart/Documents/neuromapp/b/neuromapp/keyvalue/MPI_Exec_kv  -b " << backend
     		<< " " << (async ? "-a" : "") << " " << (flash ? "-f" : "") << " -uc " << uc << " -st " << st
     		<< " -md " << md << " -dt " << dt << " -cg " << cg;
