+++ neuromapp/keyvalue/map/map_store.cpp
@@ -19,10 +19,12 @@ keyvalue_map::keyvalue_map(bool threadSafe, std::string pdsName):
                                     _numReaders(0), _numWriters(0), _async(false){
                                    _rank = mapp::master.rank();
     if (threadSafe) {
-        //		_nRdLock = new MyOMPLock();
-        //		_nWtLock = new MyOMPLock();
-        //		_readersLock = new MyOMPLock();
-        //		_writersLock = new MyOMPLock();
+#ifdef _OPENMP
+        _nRdLock = new MyOMPLock();
+        _nWtLock = new MyOMPLock();
+        _readersLock = new MyOMPLock();
+        _writersLock = new MyOMPLock();
+#endif
     } else {
         _nRdLock = new MyDummyLock();
         _nWtLock = new MyDummyLock();
