+++ neuromapp/keyvalue/meta.h
@@ -30,7 +30,9 @@
 #include <string>
 #include <vector>
 
-#include <skv/client/skv_client.hpp>
+#ifdef SKV_STORE
+#include "keyvalue/skv/meta_skv.h"
+#endif
 
 namespace keyvalue{
 
@@ -119,62 +121,6 @@ private:
     /** the value, the pointer of the memory and  the associated size is stored in std::pair */
     value_pair vp;
 };
-
-//put ifdef in the futur
-//typedef int skv_client_cmd_ext_hdl_t; // for compilation
-class meta_skv : public meta {
-public:
-    typedef meta::value_type value_type;
-    typedef value_type* pointer;
-    typedef value_type const* const_pointer;
-
-    /** \fn meta(std::string const& s, std::pair<double*,std::size_t> const& v_pair)
-	    \brief constructor of the meta information
-	    \param s the key encapsulate the key value and its size
-	    \param v_pair the value encapsulate the value pointer and the corresping size
-     */
-    explicit meta_skv(std::string const& s = std::string(),
-                       value_pair const& p = std::make_pair((pointer)NULL,0)): meta(s,p) {
-                       h = skv_client_cmd_ext_hdl_t();
-    }
-
-    /** \fn meta(std::string const& s, double* p, std::size_t n)
-	    \brief constructor of the meta information
-	    \param s the key encapsulate the key value and its size
-	    \param p the data
-     \param n size of the data
-     */
-    explicit meta_skv(std::string const& s = std::string(),
-                  const_pointer p = (pointer) NULL,
-                  std::size_t n = 0): meta(s,p,n) {
-                  h = skv_client_cmd_ext_hdl_t();
-    }
-
-    /** \fun ins_handles() const
-     \brief get the handle read only
-     */
-    inline skv_client_cmd_ext_hdl_t const* handle() const{
-        return &h;
-    }
-
-    /** \fun ins_handles()
-     \brief get the handle write only
-     */
-    inline skv_client_cmd_ext_hdl_t* handle() {
-        return &h;
-    }
-
-    /** \fun insert_reset()
-     \brief reset insert handle
-     */
-    void reset(){
-        h = skv_client_cmd_ext_hdl_t();
-    }
-
-    private:
-        /** handle */
-        skv_client_cmd_ext_hdl_t h;
-    };
 
 } //end namespace
 #endif
