+++ neuromapp/keyvalue/skv/skv_store.h
@@ -19,8 +19,8 @@ class keyvalue_skv{
 	typedef enum { SKV_NONE, SKV_INIT, SKV_CONNECT, SKV_PDS } SKV_STATE;
 
 private:
-        bool            _async;
-        int	        _rank;
+    bool        _async;
+    int	        _rank;
 	SKV_STATE	_skvState;
 	skv_client_t	_skvClient;
 	skv_pds_id_t	_pdsId;
